Data.Generics.Basics

Description

"Scrap your boilerplate" --- Generic programming in Haskell. See http://www.cs.uu.nl/wiki/GenericProgramming/SYB. This module provides the Data class with its primitives for generic programming, which is now defined in Data.Data. Therefore this module simply re-exports Data.Data.

Documentation

module Data.Data